Launch tmux sessions with pre-built panes
I tried Tmuxinator, and couldn’t get it to even build basic layout the
way I wanted it. Stole this code from @pengwynn, and it works
fantastically well. `mx` builds a new tmux session (or attaches to the
existing one a la `tat`) with a vim pane, a foreman pane (with trailing
logs), a rails console pane, and a zsh pane.