Go back to "native" autocomplete
YouCompleteMe's fuzzy completion was great, but in order to use it with tags it
had a super-slow start-up, and polluted the console history with a terrible,
scores-of-pages long list of stdout garbage. Going back to standard Vim
autocomplete with the tab wrapper, Esc to kill the window (but not insert mode),
and tab scrolling down the list.