Fix `K` grep issue when class begins with "B"
At the commandline, Vim's `<C-r><C-w>` behavior has an annoying quirk
when you're using `incsearch`. It will assume any prior characters that
match the start of your string were meant to be there, and does so in a
case-insensitive manner. For a long while I couldn't figure out why I
was unable to use my `K` mapping to grep for certain classes.

The solution is to turn off `incsearch`, perform the Ag operation, then
renable `incsearch` and preserve my sanity.