Open fzf-vim picker in a new tab
FZF in Neovim fails with an E36 error if I try to open a finder when
there's already a horizontal split open. Long-term I'd like to rescue
from the exception and only open in a tab if there's not enough space,
but this fixes the problem for the time being.