Better diff checking from production Heroku instance
This change uses Heroku's HEROKU_SLUG_COMMIT environment variable to
track the difference between production and the local branch. The prior
version of this script used the HEAD of `production/master` to check the
diff, which would fail if production had been deployed from the Heroku
Dashbaord (not with Git) or promoted through pipelines.

This script requires that the Heroku Labs dyno metadata feature be
enabled for your app:
https://devcenter.heroku.com/articles/dyno-metadata