Consolidate vim configuration from plugins to vimrc.local
I've been having trouble keeping everything in my vim configuration
together in my head, so this change puts all the stuff that used to be
in `vim/plugins` into my `vimrc.local`. A bunch of functions are still
in  `vim/autoload`, and filetype-specific stuff is still in the
appropriate directory.