Restore .git/tags support
vim-fugitive used to put this at the front of the tagfiles so that your
local tags came before gems, etc, but that changed in https://github.com/tpope/vim-fugitive/commit/63a05a6935ec4a45551bf141089c13d5671202a1.

This change brings it back with the recommendation mentioned in that
commit.