Use Yubikey for MFA with AWS CLI
The AWS CLI ordinarily has a somewhat un-ergonomic way of integrating
MFA into authentication and role assumption. aws-vault is much better at
this than the naked `aws` CLI program and can use a hardware Yubikey to
boot.

This change adds an environment variable that tells aws-vault that we're
using ykman and a Yubikey as our default MFA mechanism.